now amort list is per user, removed some console logs

+++ backend/backend/main.py
@@ -1,12 +1,14 @@
 import os
 from dotenv import load_dotenv
 
-from fastapi import FastAPI
+from fastapi import FastAPI, Depends, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 
 from supabase import create_client, Client
 
 from models import AmortEntry, BusinessProfile
+
+from auth import get_current_user
 
 # Loading env vars
 load_dotenv()
@@ -15,8 +17,8 @@ load_dotenv()
 app = FastAPI()
 
 # Supabase setup
-url = os.getenv("VITE_SUPABASE_URL")
-anon_key = os.getenv("VITE_SUPABASE_KEY")
+url = os.getenv("SUPABASE_URL")
+anon_key = os.getenv("SUPABASE_KEY")
 supabase: Client = create_client(url, anon_key)
 
 # Allows communication React <---> FastApi
@@ -30,26 +32,32 @@ app.add_middleware(
 
 # Read Amort List
 @app.get("/amort/amort-calc")
-async def get_all():
-    results = (supabase.table('amort').select('*').execute())
+async def get_all(user_id: str = Depends(get_current_user)):
+    results = (supabase.table('amort').select('*').eq("user_id", user_id).execute())
     return results.data
 
 # Create Amort
 @app.post("/amort/amort-calc")
-async def create_amort(entry: AmortEntry):
-    result = supabase.table("amort").insert(entry.model_dump(exclude={"amort_id", "created_at"})).execute()
-    return result.data[0]
+async def create_amort(entry: AmortEntry, user_id: str = Depends(get_current_user)):
+    try:
+        data = entry.model_dump(exclude={"amort_id", "created_at"})
+        data["user_id"] = user_id
+        result = supabase.table("amort").insert(data).execute()
+        return result.data[0]
+    except Exception as e:
+        print("POST error:", str(e))
+        raise HTTPException(status_code=500, detail=str(e))
 
 # Update Amort
 @app.put("/amort/amort-calc/{amort_id}")
-async def update_amort(amort_id: int, entry: AmortEntry):
-    result = supabase.table("amort").update(entry.model_dump(exclude={"amort_id", "created_at"})).eq("amort_id", amort_id).execute()
+async def update_amort(amort_id: int, entry: AmortEntry, user_id: str = Depends(get_current_user)) :
+    result = supabase.table("amort").update(entry.model_dump(exclude={"amort_id", "created_at"})).eq("amort_id", amort_id).eq("user_id", user_id).execute()
     return result.data[0]
 
 # Delete Amort
 @app.delete("/amort/amort-calc/{amort_id}")
-async def delete_amort(amort_id: int):
-    supabase.table("amort").delete().eq("amort_id", amort_id).execute()
+async def delete_amort(amort_id: int, user_id: str = Depends(get_current_user)):
+    supabase.table("amort").delete().eq("amort_id", amort_id, "user_id", user_id).execute()
     return {"message": f"Amort #{amort_id} Successfully Deleted"}
